requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
Flask-CKEditor==0.4.6
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
latex2mathml==3.73.1
MarkupSafe==2.1.1
numpy==1.23.1
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
